Move another function - no need to add a wrapper at old location
