+++ NuRadioReco/modules/efieldGalacticNoiseAdder.py
@@ -1,4 +1,4 @@
-from NuRadioReco.utilities import units, ice, geometryUtilities, trace_utilities
+from NuRadioReco.utilities import units, ice, geometryUtilities, signal_processing
 from NuRadioReco.modules.base.module import register_run
 from NuRadioReco.modules.channelGalacticNoiseAdder import channelGalacticNoiseAdder, get_local_coordinates
 import numpy as np
@@ -115,8 +115,8 @@ class efieldGalacticNoiseAdder(channelGalacticNoiseAdder):
                 self.__interpolation_frequencies, np.log10(self.__noise_temperatures[:, i_pixel]), kind='quadratic')
             noise_temperature = np.power(10, temperature_interpolator(freqs[passband_filter]))
 
-            efield_amplitude = trace_utilities.get_electric_field_from_temperature(freqs[passband_filter],
-                                                                                   noise_temperature, self.solid_angle)
+            efield_amplitude = signal_processing.get_electric_field_from_temperature(
+                freqs[passband_filter], noise_temperature, self.solid_angle)
 
             # assign random phases to electric field
             noise_spectrum = np.zeros((3, freqs.shape[0]), dtype=complex)
